+++ code/src/main/kotlin/com/expediagroup/sdk/lodgingconnectivity/sandbox/property/function/CreateSandboxPropertyFun.kt
@@ -1,9 +1,16 @@
 package com.expediagroup.sdk.lodgingconnectivity.sandbox.property.function
 
 import com.expediagroup.sdk.graphql.common.GraphQLExecutor
+import com.expediagroup.sdk.graphql.model.response.RawResponse
+import com.expediagroup.sdk.graphql.model.response.Response
 import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.SandboxCreatePropertyMutation
+import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.fragment.SandboxPropertyData
 import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.type.CreatePropertyInput
-import com.expediagroup.sdk.lodgingconnectivity.sandbox.property.model.CreateSandboxPropertyResponse
+
+data class CreateSandboxPropertyResponse(
+    override val data: SandboxPropertyData,
+    override val rawResponse: RawResponse<SandboxCreatePropertyMutation.Data>,
+) : Response<SandboxPropertyData, SandboxCreatePropertyMutation.Data>
 
 fun createSandboxPropertyFun(graphQLExecutor: GraphQLExecutor, input: CreatePropertyInput): CreateSandboxPropertyResponse {
     val operation = SandboxCreatePropertyMutation(input)
