+++ code/src/main/kotlin/com/expediagroup/sdk/lodgingconnectivity/sandbox/property/function/GetSandboxPropertiesFun.kt
@@ -3,15 +3,23 @@ package com.expediagroup.sdk.lodgingconnectivity.sandbox.property.function
 import com.expediagroup.sdk.core.extension.orNullIfBlank
 import com.expediagroup.sdk.graphql.common.GraphQLExecutor
 import com.expediagroup.sdk.graphql.model.paging.PageInfo
+import com.expediagroup.sdk.graphql.model.response.PaginatedResponse
+import com.expediagroup.sdk.graphql.model.response.RawResponse
 import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.SandboxPropertiesQuery
-import com.expediagroup.sdk.lodgingconnectivity.sandbox.property.model.SandboxPropertiesResponse
+import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.fragment.SandboxPropertyData
+
+data class GetSandboxPropertiesResponse(
+    override val data: List<SandboxPropertyData>,
+    override val rawResponse: RawResponse<SandboxPropertiesQuery.Data>,
+    override val pageInfo: PageInfo
+) : PaginatedResponse<List<SandboxPropertyData>, SandboxPropertiesQuery.Data>
 
 @JvmOverloads
 fun getSandboxPropertiesFun(
     graphQLExecutor: GraphQLExecutor,
     cursor: String? = null,
     pageSize: Int? = null
-): SandboxPropertiesResponse {
+): GetSandboxPropertiesResponse {
     val operation = SandboxPropertiesQuery
         .builder()
         .pageSize(pageSize)
@@ -30,7 +38,7 @@ fun getSandboxPropertiesFun(
         totalCount = response.data.properties.totalCount
     )
 
-    return SandboxPropertiesResponse(
+    return GetSandboxPropertiesResponse(
         data = response.data.properties.elements.map { it.sandboxPropertyData },
         rawResponse = response,
         pageInfo = currentPageInfo
