--- code/src/main/kotlin/com/expediagroup/sdk/lodgingconnectivity/sandbox/property/model/DeleteSandboxPropertyResponse.kt
@@ -1,10 +1,0 @@
-package com.expediagroup.sdk.lodgingconnectivity.sandbox.property.model
-
-import com.expediagroup.sdk.graphql.model.response.RawResponse
-import com.expediagroup.sdk.graphql.model.response.Response
-import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.SandboxDeletePropertyMutation
-
-data class DeleteSandboxPropertyResponse(
-    override val data: SandboxDeletePropertyMutation.DeleteProperty,
-    override val rawResponse: RawResponse<SandboxDeletePropertyMutation.Data>,
-) : Response<SandboxDeletePropertyMutation.DeleteProperty, SandboxDeletePropertyMutation.Data>
