+++ code/src/main/kotlin/com/expediagroup/sdk/lodgingconnectivity/sandbox/reservation/function/DeleteSandboxReservationFun.kt
@@ -1,9 +1,15 @@
 package com.expediagroup.sdk.lodgingconnectivity.sandbox.reservation.function
 
 import com.expediagroup.sdk.graphql.common.GraphQLExecutor
+import com.expediagroup.sdk.graphql.model.response.RawResponse
+import com.expediagroup.sdk.graphql.model.response.Response
 import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.SandboxDeleteReservationMutation
 import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.type.DeleteReservationInput
-import com.expediagroup.sdk.lodgingconnectivity.sandbox.reservation.model.DeleteSandboxReservationResponse
+
+data class DeleteSandboxReservationResponse(
+    override val data: SandboxDeleteReservationMutation.DeleteReservation,
+    override val rawResponse: RawResponse<SandboxDeleteReservationMutation.Data>,
+) : Response<SandboxDeleteReservationMutation.DeleteReservation, SandboxDeleteReservationMutation.Data>
 
 fun deleteSandboxReservationFun(graphQLExecutor: GraphQLExecutor, input: DeleteReservationInput): DeleteSandboxReservationResponse {
     val operation = SandboxDeleteReservationMutation(input)
