+++ code/src/main/kotlin/com/expediagroup/sdk/lodgingconnectivity/sandbox/reservation/function/GetSandboxReservationsFun.kt
@@ -3,8 +3,16 @@ package com.expediagroup.sdk.lodgingconnectivity.sandbox.reservation.function
 import com.expediagroup.sdk.core.extension.orNullIfBlank
 import com.expediagroup.sdk.graphql.common.GraphQLExecutor
 import com.expediagroup.sdk.graphql.model.paging.PageInfo
+import com.expediagroup.sdk.graphql.model.response.PaginatedResponse
+import com.expediagroup.sdk.graphql.model.response.RawResponse
 import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.SandboxPropertyReservationsQuery
-import com.expediagroup.sdk.lodgingconnectivity.sandbox.reservation.model.GetSandboxReservationsResponse
+import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.fragment.SandboxReservationData
+
+data class GetSandboxReservationsResponse(
+    override val data: List<SandboxReservationData>,
+    override val rawResponse: RawResponse<SandboxPropertyReservationsQuery.Data>,
+    override val pageInfo: PageInfo
+) : PaginatedResponse<List<SandboxReservationData>, SandboxPropertyReservationsQuery.Data>
 
 @JvmOverloads
 fun getSandboxReservations(
